@@ -3,23 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.10.2",
+    "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.9.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "bootstrap": "^4.5.0",
     "headroom.js": "^0.11.0",
+    "material-ui-image": "^3.3.1",
     "node-sass": "^4.14.1",
-    "react": "^16.13.1",
+    "react": "^16.14.0",
     "react-bootstrap": "^1.0.1",
     "react-dom": "^16.13.1",
     "react-intersection-observer": "^8.26.2",
     "react-lazyload": "^3.0.0",
+    "react-parallax": "^3.2.1",
+    "react-particles-js": "^3.4.1",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-router-hash-link": "^1.2.2",
     "react-scripts": "^3.4.3",
+    "react-scroll-trigger": "^0.6.10",
     "react-spring": "^8.0.27",
     "react-svg": "^11.0.26"
   },
