@@ -6,8 +6,6 @@ import './footer.css'
 const styles = makeStyles({
     footerStyle: {
         textAlign: 'center',
-        backgroundColor: '#f5f5f514',
-        marginTop: 'calc(5% +30px)',
     }
 })
 
@@ -16,9 +14,9 @@ const Footer = props => {
     const style = styles()
     return (
         <Fade in={true} timeout={props.timeout ? props.timeout : 1800}>
-            <Container className={`${style.footerStyle}`}>
-                <b>Copyright 2020, Mishuk Dutta, MIT license</b>
-                </Container>
+            <Container  className={style.footerStyle}>
+                <b>Copyright 2021, Mishuk Dutta, MIT license</b>
+            </Container>
         </Fade>
     )
 }
