@@ -2,13 +2,21 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
+import { Header } from './components/navBar/navBar'
+import { theme } from './components/assets/CustomTheme'
+import { ThemeProvider } from '@material-ui/core';
 import * as serviceWorker from './serviceWorker';
+import { BGParticles } from '../src/components/canvas/particles/Particles'
 
 ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+	<React.StrictMode>
+		<ThemeProvider theme={theme}>
+			<Header />
+			<App />
+		</ThemeProvider>
+		<BGParticles />
+	</React.StrictMode>,
+	document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
